lambda/frontend/requirements.txt
flask
Pillow
flask_expects_json
requests
flask-lambda-python36
http-prompt